third_party/blink/web_tests/flag-specific/disable-layout-ng/external/wpt/html/rendering/non-replaced-elements/the-fieldset-and-legend-elements/fieldset-percentage-block-size-expected.txt

This is a testharness.js-based test.
PASS rendered legend (writing-mode: horizontal-tb)
FAIL second legend (writing-mode: horizontal-tb) assert_equals: expected "30px" but got "100px"
FAIL div (writing-mode: horizontal-tb) assert_equals: expected "30px" but got "100px"
PASS rendered legend (writing-mode: vertical-lr)
FAIL second legend (writing-mode: vertical-lr) assert_equals: expected "30px" but got "100px"
FAIL div (writing-mode: vertical-lr) assert_equals: expected "30px" but got "100px"
PASS rendered legend (writing-mode: vertical-rl)
FAIL second legend (writing-mode: vertical-rl) assert_equals: expected "30px" but got "100px"
FAIL div (writing-mode: vertical-rl) assert_equals: expected "30px" but got "100px"
PASS Fieldset with a percentage height should not increase the height on every reflow.
Harness: the test ran to completion.
